Rearrange methods to avoid import dependencies
* Any attempt to access config vars from db/__init__.py causes havoc
  with the tests. It is possible to rearrange the imports in tests
  to fix this issue but this means majority of tests will need to
  setup config before any other line of code. To avoid this dependency
  move the retry method out into a standalone module yielding a lower
  chance of appearing in import orders.
